comment del bug fixed. Moved the handle comment toggle to comment componant as well as commentMenu & setCommentDelte states
