+++ client/src/components/SearchBar/SearchBar.js
@@ -97,8 +97,6 @@ export default function(props) {
         document.querySelector('.searchBar').value = '';
     };
 
-    useEffect(()=>{console.log(typeof searchResults, searchResults.length)}, [searchResults])
-
     return (
 
        <>
@@ -123,7 +121,6 @@ export default function(props) {
                 className="material-icons"
                 >search</span>
             </form>
-            {/* {searchResults.length ? */}
                 <SearchReturn
                 badSearch={badSearch}
                 setBadSearch={setBadSearch}
@@ -136,7 +133,6 @@ export default function(props) {
                 searchResults={searchResults}
                 user_id={props.user_id}
                 />
-            {/* : null} */}
         </>
     );
 };
